game0/requirements.txt
Package   Version
--------- -------
pip       25.3
pygame-ce 2.5.7
